backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
qrcode[pil]==8.0
python-dotenv==1.1.0
